// src/common/exceptions/response-exceptions.enum.ts
export enum ResponseExceptionsEnum {
  RESOURCE_NOT_FOUND = "RESOURCE_NOT_FOUND",
  RESOURCE_ALREADY_EXISTS = "RESOURCE_ALREADY_EXISTS",
  UNAUTHORIZED = "UNAUTHORIZED",
  INVALID_INPUT = "INVALID_INPUT",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  UNPROCESSABLE_ENTITY = "UNPROCESSABLE_ENTITY",
  DUE_DATE_LESS_THAN_DISPATCH_DATE = "DUE_DATE_LESS_THAN_DISPATCH_DATE",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  ORDER_NOT_PENDING = "ORDER_NOT_PENDING",
  PATIENT_HAS_RELATED_ORDERS = "PATIENT_HAS_RELATED_ORDERS",
  DENTIST_HAS_RELATED_ORDERS = "DENTIST_HAS_RELATED_ORDERS",
  SERVICE_HAS_RELATED_ORDERS = "SERVICE_HAS_RELATED_ORDERS",
  ORDER_NOT_COMPLETED = "ORDER_NOT_COMPLETED",
  ORDER_ALREADY_IN_REPORT = "ORDER_ALREADY_IN_REPORT",
}